Voice To Voice/requirements.txt
audio-recorder-streamlit
streamlit
fastapi
uvicorn
python-dotenv
datetime
openai
pathlib
requests
python-multipart